'Consulting firm's' changes

+++ create-docker-tools.sh
@@ -3,18 +3,24 @@
 ## Copy docker-tools-template to docker-tools
 ## but keep old ssh-key sub folder...
 
-chmod 777 ./docker-tools-template/start.sh
-chmod 777 ./docker-tools-template/stop.sh
+chmod 744 ./docker-tools-template/start.sh
+chmod 744 ./docker-tools-template/stop.sh
 
-rm -r docker-tools
+[ -d docker-tools/copy-to-docker-container/ssh-key ] \
+&& cp -rf docker-tools/copy-to-docker-container/ssh-key ssh-key
+
+[ -d ssh-key ] && rm -r docker-tools
+
 cp -r docker-tools-template docker-tools
-cp -r ssh-key docker-tools/copy-to-docker-container/
+
+[ -d ssh-key ] \
+&& cp -r ssh-key docker-tools/copy-to-docker-container/
 
 echo "./docker-tools/start.sh" > start
 echo "./docker-tools/stop.sh" > stop
 
-chmod 777 start
-chmod 777 stop
+chmod 744 start
+chmod 744 stop
 
 echo ""
 echo "Docker tools created! :)"
